chore(demo): move all demos into a single folder
fixes #624
