--- demo/todo/pubspec.yaml
@@ -1,7 +1,0 @@
-name: angular_dart_demo
-version: 0.0.1
-dependencies:
-  angular:
-    path: ../..
-  browser: any
-  unittest: any
