+++ demo/web/animation.html
@@ -2,7 +2,7 @@
 <html>
 <head>
     <title>NgAnimate | Demos, Stress Tests, Examples and More!</title>
-    <link type="text/css" rel="stylesheet" href="style.css" />
+    <link type="text/css" rel="stylesheet" href="css/animation.css" />
 </head>
 <body class="ng-cloak" ng-app animation-demo>
   <nav>
@@ -33,9 +33,6 @@
     </div>
     <div class="demo" ng-switch-when="Css">
       <h2>Css Demo</h2>
-      <p><strong>TODO</strong> This should contain a demo of css animation by applying multiple
-      classes and running multiple simultaneous animations on the same
-      object.</p>
       <css-demo></css-demo>
     </div>
     <div class="demo" ng-switch-when="Stress Test">
@@ -43,7 +40,7 @@
       <stress-demo></stress-demo>
     </div>
   </div>
-<script type="application/dart" src="animation_demo.dart"></script>
-<script src="packages/browser/dart.js"></script>
+<script type="application/dart" src="animation.dart"></script>
+<script src="../packages/browser/dart.js"></script>
 </body>
 </html>
