+++ demo/web/animation/repeat_demo.dart
@@ -1,4 +1,4 @@
-part of animate_demo;
+part of animation;
 
 @NgComponent(
     selector: 'repeat-demo',
