+++ demo/web/animation/stress_demo.dart
@@ -1,4 +1,4 @@
-part of animate_demo;
+part of animation;
 
 @NgComponent(
     selector: 'stress-demo',
