+++ demo/web/bouncing_balls.html
@@ -65,7 +65,7 @@
     <a href ng-click="bounce.timeDigest()">noop</a><br>
   </div>
 
-  <script type="application/dart" src="bouncy_balls.dart"></script>
+  <script type="application/dart" src="bouncing_balls.dart"></script>
   <script src="packages/browser/dart.js"></script>
 </body>
 </html>
