+++ demo/web/hello_world.html
@@ -8,7 +8,7 @@
 <h3>Hello {{ctrl.name}}!</h3>
 name: <input type="text" ng-model="ctrl.name">
 
-<script type="application/dart" src="helloworld.dart"></script>
+<script type="application/dart" src="hello_world.dart"></script>
 <script src="packages/browser/dart.js"></script>
 </body>
 </html>
