+++ demo/web/todo.html
@@ -1,12 +1,11 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <link rel="stylesheet" href="bootstrap.css">
-    <link rel="stylesheet" href="todo.css">
+    <link rel="stylesheet" href="css/bootstrap.css">
+    <link rel="stylesheet" href="css/todo.css">
     <title>Things To Do</title>
-    <script src="main.dart" type="application/dart"></script>
-    <script src="packages/browser/dart.js"></script>
-    <script src="packages/browser/interop.js"></script>
+    <script src="todo.dart" type="application/dart"></script>
+    <script src="../packages/browser/dart.js"></script>
 </head>
 <body ng-app>
